rassil/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
APScheduler==3.10.4
Pillow==10.1.0
pandas==2.1.4
openpyxl==3.1.2


